Add Packages and edit tests
