--- src/main/java/TaskType.java
@@ -1,5 +1,0 @@
-public enum TaskType {
-    T, // ToDo
-    D, // Deadline
-    E  // Event
-}
