+++ src/main/java/duke/Storage.java
@@ -1,8 +1,13 @@
+package duke;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
+
+import duke.task.Task;
+import duke.task.TasksList;
 
 public class Storage {
     private final Path path;
