+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,7 @@
+package duke.command;
+
+import duke.DukeException;
+
 public abstract class Command {
     public abstract void execute() throws DukeException;
 }
