+++ src/main/java/duke/command/UnmarkCommand.java
@@ -1,9 +1,15 @@
-public class MarkCommand extends Command {
+package duke.command;
+
+import duke.task.TasksList;
+import duke.Storage;
+import duke.DukeException;
+
+public class UnmarkCommand extends Command {
     private TasksList tasksList;
     private Storage storage;
     private String[] words;
 
-    public MarkCommand(TasksList tasksList, Storage storage, String[] words) {
+    public UnmarkCommand(TasksList tasksList, Storage storage, String[] words) {
         this.tasksList = tasksList;
         this.storage = storage;
         this.words = words;
@@ -11,10 +17,10 @@ public class MarkCommand extends Command {
 
     @Override
     public void execute() throws DukeException {
-        this.tasksList.markTask(words, storage);
+        tasksList.unmarkTask(words, storage);
     }
 
     public static boolean isCommand(String s) {
-        return s.equals("mark");
+        return s.equals("unmark");
     }
 }
