+++ src/main/java/duke/parser/Parser.java
@@ -1,3 +1,18 @@
+package duke.parser;
+
+import duke.command.Command;
+import duke.command.AddCommand;
+import duke.command.DeleteCommand;
+import duke.command.ExitCommand;
+import duke.command.InvalidCommand;
+import duke.command.ListCommand;
+import duke.command.MarkCommand;
+import duke.command.UnmarkCommand;
+import duke.task.TasksList;
+import duke.Storage;
+import duke.DukeException;
+import duke.Ui;
+
 public class Parser {
     public static Command parse(String command, Storage storage, TasksList tasksList, Ui ui) throws DukeException {
         String[] words = command.split(" ", 2);
