+++ src/main/java/duke/task/Deadline.java
@@ -1,5 +1,9 @@
+package duke.task;
+
 import java.time.DateTimeException;
 import java.time.LocalDateTime;
+
+import duke.parser.DateParser;
 
 public class Deadline extends Task {
     private LocalDateTime date;
