+++ src/main/java/duke/task/TaskType.java
@@ -1,0 +1,7 @@
+package duke.task;
+
+public enum TaskType {
+    T, // duke.task.ToDo
+    D, // duke.task.Deadline
+    E  // duke.task.Event
+}
